refactor(ToggleButton): Refactor ToggleButton to TS
