+++ packages/forma-36-react-components/src/components/IconButton/index.ts
